requirements.txt
# Core Docling distributed processing dependencies
docling>=0.1.0
nats-py>=2.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# S3 and cloud storage
aioboto3>=12.0.0
boto3>=1.34.0
botocore>=1.34.0

# HTTP client for services
httpx>=0.24.0

# Testing framework and utilities
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
pytest-mock>=3.12.0

# PDF generation for tests
reportlab>=4.0.0

# S3 mocking for tests
moto[s3]>=4.2.0

# Optional: Additional docling dependencies that might be needed
# torchvision  # For advanced image processing
# transformers # For ML models 